# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My RestAPI",
        "version": "0.1"
    },
    "paths": {
...
        "/api/v1/candidates/{id}": {
            "get": {
                "tags": [
                    "Candidates"
                ],
                "summary": "Get a specific candidate",
                "operationId": "47c34a93f02a8737f0af7f1b31a276f1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the candidate to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "data": [
                                        {
                                            "id": 1,
                                            "first_name": "ovi",
                                            "last_name": "balaji",
                                            "age": 25,
                                            "department": "computer",
                                            "min_salary_expectation": "300000",
                                            "max_salary_expectation": "600000",
                                            "code": "USD",
                                            "street_address": "Chennai,TamilNadu",
                                            "city": "Madurai",
                                            "state": "Tamil Nadu",
                                            "postal_code": "62521",
                                            "type": "mobile",
                                            "number": "910101010",
                                            "school": "Thaai School",
                                            "degree": "BE",
                                            "major": "MBA",
                                            "skill": "Software Developer",
                                            "company": "Purplematics",
                                            "title": "Project Lead",
                                            "years": 5
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Candidate not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Candidates"
                ],
                "summary": "Update a specific candidate",
                "operationId": "a79684f03c8a5ff94f7032e2bdcc0ef6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the candidate to be updated",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "age": {
                                        "type": "integer"
                                    },
                                    "department": {
                                        "type": "string"
                                    },
                                    "min_salary_expectation": {
                                        "type": "string"
                                    },
                                    "max_salary_expectation": {
                                        "type": "string"
                                    },
                                    "code": {
                                        "type": "string"
                                    },
                                    "country": {
                                        "type": "string"
                                    },
                                    "street_address": {
                                        "type": "string"
                                    },
                                    "city": {
                                        "type": "string"
                                    },
                                    "state": {
                                        "type": "string"
                                    },
                                    "postal_code": {
                                        "type": "integer"
                                    },
                                    "type": {
                                        "type": "string"
                                    },
                                    "number": {
                                        "type": "integer"
                                    },
                                    "school": {
                                        "type": "string"
                                    },
                                    "degree": {
                                        "type": "string"
                                    },
                                    "major": {
                                        "type": "string"
                                    },
                                    "skill": {
                                        "type": "string"
                                    },
                                    "level": {
                                        "type": "string"
                                    },
                                    "company": {
                                        "type": "string"
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "years": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Candidate updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Candidate updated successfully",
                                    "Candidate ID": "10"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Candidates"
                ],
                "summary": "Delete a specific candidate",
                "operationId": "6a475e6995c48a192a75d7f340c03c5d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the candidate to be deleted",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Candidate deleted successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Candidate not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/candidates": {
            "get": {
                "tags": [
                    "Candidates"
                ],
                "summary": "Get all candidates with pagination",
                "operationId": "e7ab9f74d1d3ddb6b78f785fe0559d6f",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "data": [
                                        {
                                            "id": 1,
                                            "first_name": "ovi",
                                            "last_name": "balaji",
                                            "age": 25,
                                            "department": "computer",
                                            "min_salary_expectation": "300000",
                                            "max_salary_expectation": "600000",
                                            "code": "USD",
                                            "street_address": "Chennai,TamilNadu",
                                            "city": "Madurai",
                                            "state": "Tamil Nadu",
                                            "postal_code": "62521",
                                            "type": "mobile",
                                            "number": "910101010",
                                            "school": "Thaai School",
                                            "degree": "BE",
                                            "major": "MBA",
                                            "skill": "Software Developer",
                                            "company": "Purplematics",
                                            "title": "Project Lead",
                                            "years": 5
                                        },
                                        {
                                            "id": 2,
                                            "first_name": "John",
                                            "last_name": "Cena",
                                            "age": 40,
                                            "department": "Tester",
                                            "min_salary_expectation": "900000",
                                            "max_salary_expectation": "900000",
                                            "code": "USD",
                                            "street_address": "Samuthar,Delhi",
                                            "city": "Red Fort",
                                            "state": "Delhi",
                                            "postal_code": "192033",
                                            "type": "mobile",
                                            "number": "9030101010",
                                            "school": "Gandhi School",
                                            "degree": "BE",
                                            "major": "MBA",
                                            "skill": "Software Tester",
                                            "company": "TCS",
                                            "title": "Project Lead",
                                            "years": 10
                                        }
                                    ],
                                    "meta": {
                                        "current_page": 1,
                                        "per_page": 10,
                                        "total": 20
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Candidates"
                ],
                "summary": "Create a new candidate",
                "operationId": "cde870ee041ab3c846c9e4bd63929e15",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "age": {
                                        "type": "integer"
                                    },
                                    "department": {
                                        "type": "string"
                                    },
                                    "min_salary_expectation": {
                                        "type": "string"
                                    },
                                    "max_salary_expectation": {
                                        "type": "string"
                                    },
                                    "code": {
                                        "type": "string"
                                    },
                                    "country": {
                                        "type": "string"
                                    },
                                    "street_address": {
                                        "type": "string"
                                    },
                                    "city": {
                                        "type": "string"
                                    },
                                    "state": {
                                        "type": "string"
                                    },
                                    "postal_code": {
                                        "type": "integer"
                                    },
                                    "type": {
                                        "type": "string"
                                    },
                                    "number": {
                                        "type": "integer"
                                    },
                                    "school": {
                                        "type": "string"
                                    },
                                    "degree": {
                                        "type": "string"
                                    },
                                    "major": {
                                        "type": "string"
                                    },
                                    "skill": {
                                        "type": "string"
                                    },
                                    "level": {
                                        "type": "string"
                                    },
                                    "company": {
                                        "type": "string"
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "years": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Candidate created successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Candidate created successfully",
                                    "Candidate ID": "10"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}